+++ src/main/java/jpa/domain/repository/FavoriteRepository.java
@@ -1,6 +1,8 @@
-package jpa.domain;
+package jpa.domain.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+
+import jpa.domain.entity.Favorite;
 
 public interface FavoriteRepository extends JpaRepository<Favorite, Long> {
 }
