+++ src/main/java/jpa/domain/repository/LineRepository.java
@@ -1,6 +1,8 @@
-package jpa.domain;
+package jpa.domain.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+
+import jpa.domain.entity.Line;
 
 public interface LineRepository extends JpaRepository<Line, Long> {
 	Line findByName(String name);
