+++ src/main/java/jpa/domain/repository/MemberRepository.java
@@ -1,6 +1,8 @@
-package jpa.domain;
+package jpa.domain.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+
+import jpa.domain.entity.Member;
 
 public interface MemberRepository extends JpaRepository<Member, Long> {
 }
