+++ src/main/java/jpa/domain/repository/StationRepository.java
@@ -1,6 +1,8 @@
-package jpa.domain;
+package jpa.domain.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+
+import jpa.domain.entity.Station;
 
 public interface StationRepository extends JpaRepository<Station, Long> {
 	Station findByName(String name);
